Refactor: error handlers
